I have been trying to use litex-buildenv to make a vertexriscv image for the HX8K. I am using the stub firmware.
The built firmware does not come up.

We tried adding the UARTWishboneBridge, but `wishbone-tool` cannot connect to it (as if the Wishbone bus is locking up).

Moving the SRAM to address 0 makes it possible to inspect the wishbone bus with the tool.

The CPU was not running code correctly, changing the SPIFlash to little endian corrects this.

At this point, with UARTWishboneBridge still in, the CPU starts running the BIOS but is stuck in an infinite loop apparently in a call to vsnprintf in the BIOS that I cannot locate.

Removing the UARTWishboneBridge does not make the BIOS come up.